requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.24.1
pydantic[email]==2.5.0
supabase==1.2.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
python-multipart==0.0.6
email-validator==2.1.0
cryptography==41.0.7
resend==0.8.0
psycopg2-binary==2.9.9
redis==5.0.0
tabulate==0.9.0